backend/app/core/security.py
import uuid
from datetime import datetime, timedelta, timezone

from jose import jwt

from app.config import settings


def create_token(user_id: uuid.UUID, role: str, status: str, expires_delta: timedelta) -> str:
    now = datetime.now(timezone.utc)
    payload = {
        "sub": str(user_id),
        "role": role,
        "status": status,
        "jti": str(uuid.uuid4()),
        "iat": now,
        "exp": now + expires_delta,
    }
    return jwt.encode(payload, settings.secret_key, algorithm=settings.algorithm)


def create_access_token(user_id: uuid.UUID, role: str, status: str) -> str:
    return create_token(
        user_id, role, status, timedelta(minutes=settings.access_token_expire_minutes)
    )


def create_refresh_token(user_id: uuid.UUID, role: str, status: str) -> str:
    return create_token(
        user_id, role, status, timedelta(days=settings.refresh_token_expire_days)
    )


def decode_token(token: str) -> dict:
    return jwt.decode(token, settings.secret_key, algorithms=[settings.algorithm])
